requirements.txt
numpy
PyQt5
numexpr
seaborn
Pillow
opencv-python
matplotlib
pandas
typing_extensions
pyinstaller
pyvista